requirements.txt
streamlit
pandas
numpy
scikit-learn
seaborn
matplotlib
joblib
reportlab
streamlit
gspread
google-auth

